EVG-17375 Create manifest upon patch finalization
